Use vitest instead of mocha
